Replace vue-i18n with i18next
i18n has more features, has better docs,
is more straightforward and is more widely used.
Features that let you do more without functions in the lang files.
Still uses the $t('lang_key') syntax with a plugin for vue so no
changes there. Also reverted to using a lang import file as the
available vite plugin doesn't work, we don't have thaaat many translations,
and if we do have too many in the future, it's easier to work from
for partial imports.
json5 can be imported but the ts checker doesn't like it so I added
the ts-nocheck for now.....
There is a compiler option allowArbitraryExtensions, but it's
not working on my ver of the compiler yet.
I'd like to keep the files as json5 just so that they're language
independent, simply key and value strings, so that they can be
e.g. put in a database or possibly read by some other software
for easier translating and keeping in sync in the future.